requirements.txt
streamlit==1.32.0
opencv-python==4.8.0.76
numpy==1.24.4
torch==2.1.0
torchvision==0.16.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.8.2
scikit-learn==1.3.2
mediapipe==0.10.8
seaborn==0.13.2
reportlab==4.0.9
